[misc] code format standardization following google-java-format

+++ src/main/java/org/mariadb/jdbc/internal/util/pid/PidFactory.java
@@ -52,33 +52,33 @@
 
 package org.mariadb.jdbc.internal.util.pid;
 
-import java.lang.reflect.Method;
-import java.util.function.Supplier;
+import java.lang.reflect.*;
+import java.util.function.*;
 
 public class PidFactory {
   private static Supplier<String> instance;
 
   static {
     try {
-      //if java 9+
+      // if java 9+
       Class<?> processHandle = Class.forName("java.lang.ProcessHandle");
-      instance = () -> {
-        try {
-          Method currentProcessMethod = processHandle.getMethod("current");
-          Object currentProcess = currentProcessMethod.invoke(null);
-          Method pidMethod = processHandle.getMethod("pid");
-          return String.valueOf((Long) pidMethod.invoke(currentProcess));
-        } catch (Throwable throwable) {
-          return null;
-        }
-      };
+      instance =
+          () -> {
+            try {
+              Method currentProcessMethod = processHandle.getMethod("current");
+              Object currentProcess = currentProcessMethod.invoke(null);
+              Method pidMethod = processHandle.getMethod("pid");
+              return String.valueOf(pidMethod.invoke(currentProcess));
+            } catch (Throwable throwable) {
+              return null;
+            }
+          };
     } catch (Throwable cle) {
       try {
         instance = JnaPidFactory.getInstance();
       } catch (Throwable throwable) {
         instance = () -> null;
       }
-
     }
   }
 
@@ -86,4 +86,3 @@ public class PidFactory {
     return instance;
   }
 }
-
